GH-29: :building_construction: :truck: Rename `AppointmentScheduleSearch` to `AppointmentScheduleSearchForm` and refactor implementation
