+++ app/screens/appointmentSchedule/AppointmentScheduleSearchForm.stories.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import { storiesOf } from '@storybook/react';
+
+import AppointmentScheduleSearchForm from './AppointmentScheduleSearchForm';
+
+storiesOf(
+  'Screens/AppointmentSchedule/AppointmentScheduleSearchForm',
+  module
+).add('default', () => <AppointmentScheduleSearchForm />);
